project/requirements.txt
asgiref==3.2.10
pytz==2020.1
sqlparse==0.2.4

Django>=3.0,<3.1
djangorestframework==3.11.1

python-decouple
celery>=4.4, <4.5
django-celery-beat
drf-yasg==1.17.1
sentry-sdk==0.19.1

psycopg2-binary
zeep
coverage==5.1
mock==4.0.2
django-filter
django-json-widget==1.0.1
coreapi==2.3.3
